// src/types/Banner.ts
export interface Banner {
  uuid: string;
  title_uz: string;
  title_kr: string;
  title_ru: string;
  title_en: string;
  title_tj: string;
  title_kz: string;
  title_kg: string;
  subtitle_uz: string;
  subtitle_kr: string;
  subtitle_ru: string;
  subtitle_en: string;
  subtitle_tj: string;
  subtitle_kz: string;
  subtitle_kg: string;
  text_uz: string;
  text_kr: string;
  text_ru: string;
  text_en: string;
  text_tj: string;
  text_kz: string;
  text_kg: string;
  order: number;
  image: string;
  video?: string;
}